Move Dockerfile and fix project-compose.yml:
The Dockerfile rules COPY and ADD cannot access files outside
the Dockerfile context (`cd ..`), because this, it was moved
backward and now it can access the "./src/" content.

The version field of the `project-compose.yml` was updated
(from 3.1) to 3.6, a simple fix was done in "app-0" and the
Dockefile build context was updated.